requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
black==21.5b1
click==7.1.2
Django==3.2.2
djangorestframework==3.12.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pylint==2.8.2
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
